refactor: move FFT to viewmodel package
